Remove asString() method from TextFileContent
